+++ bookchain/asyncql/Account.py
@@ -1,12 +1,11 @@
 from __future__ import annotations
+from .Entry import Entry
+from bookchain.enums import AccountType, EntryType
 from sqloquent.asyncql import (
     AsyncHashedModel, AsyncRelatedModel, AsyncRelatedCollection,
     AsyncQueryBuilderProtocol, Default
 )
 from tapescript import run_auth_scripts, Script
-from .AccountType import AccountType
-from .Entry import Entry
-from .EntryType import EntryType
 import packify
 
 
@@ -113,6 +112,7 @@ class Account(AsyncHashedModel):
     @classmethod
     def query(cls, conditions: dict = None, connection_info: str = None) -> AsyncQueryBuilderProtocol:
         """Ensure conditions are encoded before querying."""
+        conditions = {**conditions} if conditions else {}
         if conditions and type(conditions.get('type', None)) is AccountType:
             conditions['type'] = conditions['type'].value
         return super().query(conditions, connection_info)