+++ bookchain/models/EntryType.py
@@ -1,7 +1,0 @@
-from enum import Enum
-
-
-class EntryType(Enum):
-    """Enum of valid Entry types (CREDIT and DEBIT)."""
-    CREDIT = 'c'
-    DEBIT = 'd'