+++ bookchain/models/LedgerType.py
@@ -1,9 +1,0 @@
-from enum import Enum
-
-
-class LedgerType(Enum):
-    """Enum of valid ledger types: CURRENT and FUTURE for cash and
-        accrual accounting, respectively.
-    """
-    CURRENT = 'Current'
-    FUTURE = 'Future'